now shared llm core
